// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleName = {
  ADMIN: 'ADMIN',
  SUBADMIN: 'SUBADMIN',
  USER: 'USER'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const OrderStatus = {
  PENDING: 'PENDING',
  AWAITING_PAYMENT: 'AWAITING_PAYMENT',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const CartItemStatus = {
  ACTIVE: 'ACTIVE',
  REMOVED: 'REMOVED'
} as const

export type CartItemStatus = (typeof CartItemStatus)[keyof typeof CartItemStatus]